+++ micro-game-lobby/src/lib.rs
@@ -1,14 +1,9 @@
-extern crate bincode;
-extern crate community_redis_streams;
-extern crate core_model;
-extern crate crossbeam_channel;
-extern crate log;
-extern crate redis;
-
-pub mod components;
+mod components;
 mod game_lobby;
 mod repo;
 pub mod stream;
 pub mod topics;
+
+pub use components::*;
 
 pub const PUBLIC_GAME_BOARD_SIZE: u16 = 19;