+++ micro-game-lobby/src/main.rs
@@ -1,11 +1,13 @@
-use micro_game_lobby::*;
 use log::info;
-use components::Components;
+use micro_game_lobby::*;
 
 const VERSION: &str = env!("CARGO_PKG_VERSION");
 
 fn main() {
     env_logger::init();
     info!("🔢 {}", VERSION);
-    stream::process(&Components::default())
+    let client = redis_client();
+    let components = Components::new(client.clone());
+    stream::create_consumer_group(&client);
+    stream::process(&components)
 }